Ajout des conditions et réorganisation du code
- Ajout de la gestion des conditions (voir wiki)
- Grosse amélioration du code (optimisation/clarté)
- Normalement plus de fuite mémoire dans les scénarios nominaux
- Ajout d'exemple

+++ src/err/err.c
@@ -10,8 +10,8 @@ int nb_alloc = 0;
 size_t size_alloc = 0;
 
 void print_memory_info(void) {
-    printf("[i] %d memory block(s) not free'd \n", nb_alloc);
-    printf("[i] %d octets malloc'd \n", size_alloc/8);
+    printf("[i] %u memory block(s) not free'd \n", nb_alloc);
+    printf("[i] %lu octets malloc'd \n", (long unsigned int)size_alloc/8);
 }
 
 void* xmalloc(size_t s) {
